+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoCode.java
@@ -1,0 +1,76 @@
+package org.firstinspires.ftc.teamcode;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.IMU;
+
+import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.DcMotorSimple;
+import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
+import com.qualcomm.robotcore.hardware.Servo;
+import com.qualcomm.robotcore.util.ElapsedTime;
+
+@Autonomous
+public class AutoCode extends LinearOpMode {
+
+    @Override
+    public void runOpMode() throws InterruptedException {
+        // Declare our motors
+        // Make sure your ID's match your configuration
+
+        DcMotor frontLeftMotor = hardwareMap.dcMotor.get("frontLeftMotor");
+        DcMotor backLeftMotor = hardwareMap.dcMotor.get("backLeftMotor");
+        DcMotor frontRightMotor = hardwareMap.dcMotor.get("frontRightMotor");
+        DcMotor backRightMotor = hardwareMap.dcMotor.get("backRightMotor");
+
+        Servo clawServo1 = hardwareMap.get(Servo.class, "ClawServo1");
+        Servo clawServo2 = hardwareMap.get(Servo.class, "ClawServo2");
+
+        clawServo1.setPosition(0.40);
+        clawServo2.setPosition(0.65);
+
+        double position = 0;
+
+        ElapsedTime timeSinceRightBumperPressed = new ElapsedTime();
+
+        boolean claw_state = false; // false as open, true as closed
+
+        //double claw1_position = 0.5;
+        //double claw2_position = 0.5;
+
+        // Reverse the right side motors. This may be wrong for your setup.
+        // If your robot moves backwards when commanded to go forwards,
+        // reverse the left side instead.
+        // See the note about this earlier on this page.
+        frontLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+        backLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+
+        // Retrieve the IMU from the hardware map
+        IMU imu = hardwareMap.get(IMU.class, "imu");
+        // Adjust the orientation parameters to match your robot
+        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
+                RevHubOrientationOnRobot.LogoFacingDirection.FORWARD,
+                RevHubOrientationOnRobot.UsbFacingDirection.UP));
+        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
+        imu.initialize(parameters);
+
+        // movement functions
+
+
+        waitForStart();
+        // for some reason does not add telemetry data idk why
+        telemetry.addData("running","True");
+
+        // goes forward with 20% power for half a second
+        frontLeftMotor.setPower(20);
+        frontRightMotor.setPower(20);
+        backLeftMotor.setPower(20);
+        backRightMotor.setPower(20);
+        sleep(500);
+        frontLeftMotor.setPower(0);
+        frontRightMotor.setPower(0);
+        backLeftMotor.setPower(0);
+        backRightMotor.setPower(0);
+
+    }
+}
