--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MecanumDriveTrain.java
@@ -1,124 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.IMU;
-
-import java.security.cert.LDAPCertStoreParameters;
-import java.util.ArrayList;
-import java.util.List;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.util.ElapsedTime;
-import java.util.ArrayList;
-
-@TeleOp
-public class MecanumDriveTrain extends LinearOpMode {
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        // Declare our motors
-        // Make sure your ID's match your configuration
-
-        DcMotor frontLeftMotor = hardwareMap.dcMotor.get("frontLeftMotor");
-        DcMotor backLeftMotor = hardwareMap.dcMotor.get("backLeftMotor");
-        DcMotor frontRightMotor = hardwareMap.dcMotor.get("frontRightMotor");
-        DcMotor backRightMotor = hardwareMap.dcMotor.get("backRightMotor");
-
-        Servo clawServo1 = hardwareMap.get(Servo.class, "ClawServo1");
-        Servo clawServo2 = hardwareMap.get(Servo.class, "ClawServo2");
-
-        clawServo1.setPosition(0.40);
-        clawServo2.setPosition(0.65);
-
-        double position = 0;
-
-        ElapsedTime timeSinceRightBumperPressed = new ElapsedTime();
-
-        boolean claw_state = false; // false as open, true as closed
-
-        //double claw1_position = 0.5;
-        //double claw2_position = 0.5;
-
-        // Reverse the right side motors. This may be wrong for your setup.
-        // If your robot moves backwards when commanded to go forwards,
-        // reverse the left side instead.
-        // See the note about this earlier on this page.
-        frontLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
-        backLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        // Retrieve the IMU from the hardware map
-        IMU imu = hardwareMap.get(IMU.class, "imu");
-        // Adjust the orientation parameters to match your robot
-        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.FORWARD,
-                RevHubOrientationOnRobot.UsbFacingDirection.UP));
-        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
-        imu.initialize(parameters);
-
-        waitForStart();
-
-
-
-        while (opModeIsActive()) {
-
-            double y = -gamepad1.left_stick_y; // Remember, Y stick value is reversed
-            double x = -gamepad1.left_stick_x;
-            double rx = gamepad1.right_stick_x;
-
-
-
-            if(gamepad1.right_bumper) {
-                if(timeSinceRightBumperPressed.milliseconds() > 500) {
-                    timeSinceRightBumperPressed.reset();
-                    claw_state = !claw_state;
-                }
-
-
-            }
-
-            if(claw_state) {
-                clawServo1.setPosition(0.05);
-                clawServo2.setPosition(1);
-            }
-            else {
-                clawServo1.setPosition(0.40);
-                clawServo2.setPosition(0.65);
-            }
-
-
-            if (gamepad1.options) {
-                imu.resetYaw();
-            }
-
-            double botHeading = -imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
-
-            // Rotate the movement direction counter to the bot's rotation
-            double rotX = x * Math.cos(-botHeading) - y * Math.sin(-botHeading);
-            double rotY = x * Math.sin(-botHeading) + y * Math.cos(-botHeading);
-
-            rotX = rotX * 1.1;  // Counteract imperfect strafing
-
-            // Denominator is the largest motor power (absolute value) or 1
-            // This ensures all the powers maintain the same ratio,
-            // but only if at least one is out of the range [-1, 1]
-            double denominator = Math.max(Math.abs(rotY) + Math.abs(rotX) + Math.abs(rx), 1);
-            double frontLeftPower = (rotY + rotX + rx) / denominator;
-            double backLeftPower = (rotY - rotX + rx) / denominator;
-            double frontRightPower = (rotY - rotX - rx) / denominator;
-            double backRightPower = (rotY + rotX - rx) / denominator;
-
-            frontLeftMotor.setPower(frontLeftPower);
-            backLeftMotor.setPower(backLeftPower);
-            frontRightMotor.setPower(frontRightPower);
-            backRightMotor.setPower(backRightPower);
-
-
-        }
-    }
-}
-
